requirements.txt
streamlit
psycopg2-binary
pandas
plotly
openai
python-dotenv
